[None][test] Fold fused-scatter parity test into the MSA backend test module
Keep only the reference-parity test (8 parametrized cases covering fp8/bf16
caches, 1/4 KV heads, with/without the index cache) and move it into
test_minimax_m3_msa_backend.py; drop the layout-rejection, int64-offset, and
boundary-slot micro-tests along with the standalone scatter test file.
